Continue centralizing utils
