+++ front/app/components/Form/utils/getFormCompletionPercentage.test.ts
@@ -1,125 +1,6 @@
-import {
-  getFilteredDataForUserPath,
-  isPageCategorization,
-  PageType,
-  getFormCompletionPercentage,
-} from 'components/Form/Components/Layouts/utils';
+import { PageType } from '../typings';
 
-describe('isPageCategorization', () => {
-  it('should return true if the categorization is a page', () => {
-    const uiSchema = {
-      type: 'Categorization',
-      label: 'Testlabel',
-      options: {
-        id: 'extra',
-      },
-      elements: [
-        {
-          type: 'Page',
-          label: 'Testlabel',
-          options: {
-            id: 'extra',
-          },
-          elements: [],
-        },
-      ],
-    };
-
-    const jsonSchema = {
-      type: 'object',
-      additionalProperties: false,
-      properties: {},
-      required: ['short_answer'],
-    };
-    const isPage = isPageCategorization(uiSchema, jsonSchema);
-    expect(isPage).toEqual(true);
-  });
-
-  it('should return false when the structure has no page in it', () => {
-    const uiSchema = {
-      type: 'Categorization',
-      label: 'Testlabel',
-      options: {
-        id: 'extra',
-      },
-      elements: [
-        {
-          type: 'Control',
-          label: 'Testlabel',
-          options: {
-            id: 'extra',
-          },
-          elements: [],
-        },
-      ],
-    };
-
-    const jsonSchema = {
-      type: 'object',
-      additionalProperties: false,
-      properties: {},
-      required: ['short_answer'],
-    };
-    const isPage = isPageCategorization(uiSchema, jsonSchema);
-    expect(isPage).toEqual(false);
-  });
-});
-
-describe('getFilteredDataForUserPath', () => {
-  it('should return data only present in the user page path', () => {
-    const pageCategorization: PageType[] = [
-      {
-        type: 'Page',
-        label: 'Testlabel',
-        options: {
-          id: 'extra',
-        },
-        elements: [
-          {
-            type: 'Control',
-            scope: '#/properties/multiple_choice',
-            label: 'Fruits',
-            options: {
-              description: '',
-              isAdminField: false,
-            },
-          },
-        ],
-      },
-      {
-        type: 'Page',
-        label: 'Testlabel',
-        options: {
-          id: 'extra',
-        },
-        elements: [
-          {
-            type: 'Control',
-            scope: '#/properties/multiple_choice_1',
-            label: 'Fruits',
-            options: {
-              description: '',
-              isAdminField: false,
-            },
-          },
-        ],
-      },
-    ];
-
-    let filteredData = getFilteredDataForUserPath(pageCategorization, {
-      multiple_choice: 'value',
-    });
-    expect(filteredData).toEqual({ data: { multiple_choice: 'value' } });
-    filteredData = getFilteredDataForUserPath(pageCategorization, {
-      multiple_choice_1: 'value',
-    });
-    expect(filteredData).toEqual({ data: { multiple_choice_1: 'value' } });
-    filteredData = getFilteredDataForUserPath(pageCategorization, {
-      multiple_choice_2: 'value',
-    });
-    expect(filteredData).toEqual({ data: {} });
-  });
-});
+import getFormCompletionPercentage from './getFormCompletionPercentage';
 
 describe('getFormCompletionPercentage', () => {
   it('returns 0 when no fields are filled in and user is on first page', () => {
