+++ front/app/components/Form/utils/isPageCategorization.ts
@@ -1,0 +1,25 @@
+import { and, isCategorization, Tester, uiTypeIs } from '@jsonforms/core';
+import { isEmpty } from 'lodash-es';
+
+import { PageCategorization } from '../typings';
+
+const isPageCategorization: Tester = and(
+  uiTypeIs('Categorization'),
+  (uischema) => {
+    const hasPage = (element: PageCategorization): boolean => {
+      if (isEmpty(element.elements)) {
+        return false;
+      }
+
+      return element.elements
+        .map((elem) =>
+          isCategorization(elem) ? hasPage(elem) : elem.type === 'Page'
+        )
+        .reduce((prev, curr) => prev && curr, true);
+    };
+
+    return hasPage(uischema as PageCategorization);
+  }
+);
+
+export default isPageCategorization;
